# Conflicts:
#	.gitignore
#	MSVC19/rheaAlipayChina/rheaAlipayChina.vcxproj
#	QT-Linaro-Rockchip/CPUBridge/CPUBridge.pro
#	QT-Linaro-Rockchip/SocketBridge/SocketBridge.pro
#	QT-Linaro-Rockchip/rheaAlipayChina/rheaAlipayChina.pro
#	QT-Linaro-Rockchip/rheaAppLib/rheaAppLib.pro
#	QT-Linaro-Rockchip/rheaCommonLib/rheaCommonLib.pro
#	QT-Linaro-Rockchip/rheaDB/rheaDB.pro
#	QT-Linaro-Rockchip/rheaDB/rheaDB.pro.user.97c4efb.4.8-pre1
#	QT5/GPUProject.pro.user
#	QT5/SMUProject.pro.user
#	QT5/rheaDB/rheaDB.pro.user
#	QT5/rheaExternalSerialAPI/rheaExternalSerialAPI.pro
#	bin/aaa_MH210_CREATE_ESAPI.sh
#	bin/aaa_MH6_CREATE_ESAPI.sh
#	bin/aaa_MH6_CREATE_rheAPI.sh
#	src/CPUBridge/CPUBridge.cpp
#	src/GPU/header.h
#	src/GPU/main.cpp
#	src/GPU/mainwindow.cpp
#	src/rheaCommonLib/Protocol/ProtocolWebsocket.cpp
#	src/rheaCommonLib/rhea.h
#	src/rheaCommonLib/rheaString.h
#	src/rheaCommonLib/string/rheaStringEnumAndDefine.h
#	src/rheaCommonLib/string/rheaUTF8.cpp
#	src/rheaExternalSerialAPI/ESAPICore.cpp
#	src/rheaExternalSerialAPI/ESAPIModuleRasPI.cpp
#	src/rheaExternalSerialAPI/ESAPIModuleRasPI.h

+++ src/rheaExternalSerialAPI/ESAPIModuleRasPI.cpp
@@ -89,7 +89,7 @@ void ModuleRasPI::virt_handleMsgFromSocket (sShared *shared, OSSocket &sok, u32 
         DBGBREAK;
         break;
 
-    case ModuleRasPI::eStato::running:  
+    case ModuleRasPI::eStato::running:
         running_handleMsgFromSocket (shared, sok, userParam);
         break;
 
@@ -269,8 +269,8 @@ void ModuleRasPI::boot_handleMsgFromSubscriber(sShared *shared, sSubscription &s
             const u8 *filename;
             const u8 *dstFolder;
             esapi::translate_RASPI_UNZIP(msg, &filename, &dstFolder);
-            const u8 len1 = rhea::string::utf8::lengthInBytes(filename);
-            const u8 len2 = rhea::string::utf8::lengthInBytes(dstFolder);
+            const u8 len1 = (u8)rhea::string::utf8::lengthInBytes(filename);
+            const u8 len2 = (u8)rhea::string::utf8::lengthInBytes(dstFolder);
 
             //chiedo al rasPI
             //# R [0x05] [lenFilename] [lenFolder] [filename_terminato_con_0x00] [folderDest_con_0x00] [ck]
@@ -584,7 +584,7 @@ void ModuleRasPI::running_handleMsg_R_fromRs232	(sShared *shared, sBuffer *b)
                 if (b->numBytesInBuffer < 10 + dataLen)
                     return;
 
-                const u8* data = &b->buffer[9];
+                const u8 *data = &b->buffer[9];
                 const u8 ck = b->buffer[9 + dataLen];
                 if (rhea::utils::simpleChecksum8_calc(b->buffer, 9 + dataLen) != ck)
                 {
@@ -598,7 +598,7 @@ void ModuleRasPI::running_handleMsg_R_fromRs232	(sShared *shared, sBuffer *b)
                     sConnectedSocket *cl = priv_2280_findConnectedSocketByUID(uid);
                     if (NULL != cl)
                     {
-                        rhea::socket::write (cl->sok, data, dataLen);
+                        rhea::socket::write (cl->sok, data, (u16)dataLen);
                         shared->logger->log ("esapi::ModuleRasPI::RS232 => rcv [%d] bytes, sending to socket [%d]\n", dataLen, cl->uid);
                     }
                     else
